apps/fcp-subs-whisper/pyproject.toml
[project]
name = "fcp-subs-whisper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faster-whisper>=1.2.1",
    "mlx-whisper>=0.4.3",
    "tqdm>=4.67.3",
    "wyoming>=1.8.0",
    "pyannote.audio>=3.3.1",
    "torch>=2.5.1",
]

[project.scripts]
fcp-subs-whisper = "fcp_subs_whisper.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/fcp_subs_whisper"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
